Move test files to index directory
Renamed and relocated test files to the test/index directory for better organization. Updated import paths in affected files to reflect new locations.
