+++ test/index/token.spec.ts
